+++ src/index.js
@@ -16,16 +16,29 @@ const pluginReleaseUrl = constants.PLUGINRELEASEURL;
 let spotifyPageName = "/TPSpotifyPG.tml"
 let updateAuto = "false";
 let updateFrequency = constants.START_CAPTURE_WAIT_TIME;
+let updateIsPlayingAuto = "false";
+let updateTrackInfoAuto = "false";
 
 const spotifyosa = require('./osascript');
 let previousIsOpenState = undefined;
 let previousIsVisibleState = undefined;
+let previousIsPlayingState = undefined;
+
+let currentTrackTitle = undefined;
+let currentTrackAlbum = undefined;
+let currentTrackArtist = undefined;
+let currentTrackAlbumArtist = undefined;
+let currentTrackDuration = undefined;
+let currentTrackId = undefined;
+let currentTrackArtworkUrl = undefined;
 
 let updateLoop = undefined;
 const settings = {
     [constants.SETTING_SPOTIFY_PAGE_NAME]: spotifyPageName,
     [constants.SETTING_AUTOMATIC_UPDATE]: updateAuto,
     [constants.SETTING_AUTOMATIC_UPDATE_FREQUENCY]: updateFrequency,
+    [constants.SETTING_IS_PLAYING_AUTOMATIC_UPDATE]: updateIsPlayingAuto,
+    [constants.SETTING_TRACK_INFO_AUTOMATIC_UPDATE]: updateTrackInfoAuto,
 };
 
 tpclient.on("Settings", (data) => {
@@ -52,6 +65,16 @@ tpclient.on("Settings", (data) => {
     if (updateFrequency != (settings[constants.SETTING_AUTOMATIC_UPDATE_FREQUENCY])) {
         updateFrequency = (settings[constants.SETTING_AUTOMATIC_UPDATE_FREQUENCY]);
         tpclient.logIt("DEBUG", "Settings: Update spotify state frequency is set to", updateFrequency, "seconds");
+    }
+
+    if (updateIsPlayingAuto != settings[constants.SETTING_IS_PLAYING_AUTOMATIC_UPDATE]) {
+        updateIsPlayingAuto = settings[constants.SETTING_IS_PLAYING_AUTOMATIC_UPDATE];
+        tpclient.logIt("DEBUG", "Settings: Update is playing state automatically is set to", updateIsPlayingAuto);
+    }
+
+    if (updateTrackInfoAuto != settings[constants.SETTING_TRACK_INFO_AUTOMATIC_UPDATE]) {
+        updateTrackInfoAuto = settings[constants.SETTING_TRACK_INFO_AUTOMATIC_UPDATE];
+        tpclient.logIt("DEBUG", "Settings: Update track info automatically is set to", updateTrackInfoAuto);
     }
 
     updateSpotifyState(true);
@@ -103,6 +126,70 @@ tpclient.on("Action", async (data) => {
                             tpclient.stateUpdate("state_spotifypg_visible", "true");
                         });
                     }
+                });
+            }
+        });
+    }
+    else if (data.actionId === "action_spotifypg_play") {
+        spotifyosa.isApplicationOpen().then((isOpen) => {
+            if (isOpen) {
+                spotifyosa.isApplicationPlaying().then((isPlaying) => {
+                    if (isPlaying == "paused") {
+                        spotifyosa.playCurrentTrack().then(() => {
+                            previousIsPlayingState = "playing";
+                            tpclient.stateUpdate("state_spotifypg_play", "playing");
+
+                            getTrackInfoStates();
+                        });
+                    }
+                });
+            }
+        });
+    }
+    else if (data.actionId === "action_spotifypg_pause") {
+        spotifyosa.isApplicationOpen().then((isOpen) => {
+            if (isOpen) {
+                spotifyosa.isApplicationPlaying().then((isPlaying) => {
+                    if (isPlaying == "playing") {
+                        spotifyosa.pauseCurrentTrack().then(() => {
+                            previousIsPlayingState = "paused";
+                            tpclient.stateUpdate("state_spotifypg_play", "paused");
+
+                            getTrackInfoStates();
+                        });
+                    }
+                });
+            }
+        });
+    }
+    else if (data.actionId === "action_spotifypg_toggle_play") {
+        spotifyosa.isApplicationOpen().then((isOpen) => {
+            if (isOpen) {
+                spotifyosa.playpauseCurrentTrack().then(() => {
+                    spotifyosa.isApplicationPlaying().then((isPlaying) => {
+                        previousIsPlayingState = isPlaying;
+                        tpclient.stateUpdate("state_spotifypg_play", isPlaying);
+
+                        getTrackInfoStates();
+                    });
+                });
+            }
+        });
+    }
+    else if (data.actionId === "action_spotifypg_next") {
+        spotifyosa.isApplicationOpen().then((isOpen) => {
+            if (isOpen) {
+                spotifyosa.playNextTrack().then(() => {
+                    getIsPlayingState(true, true);
+                });
+            }
+        });
+    }
+    else if (data.actionId === "action_spotifypg_previous") {
+        spotifyosa.isApplicationOpen().then((isOpen) => {
+            if (isOpen) {
+                spotifyosa.playPreviousTrack().then(() => {
+                    getIsPlayingState(true, true);
                 });
             }
         });
@@ -214,6 +301,7 @@ function updateSpotifyState(shouldForceUpdate) {
 
         if (isOpen) {
             getIsVisibleState();
+            getIsPlayingState(shouldForceUpdate);
         }
     });
 }
@@ -227,9 +315,77 @@ function getIsVisibleState() {
     });
 }
 
+function getIsPlayingState(shouldForceUpdate, shouldUpdateTrackInfoStates) {
+    if (updateIsPlayingAuto == "true" || shouldForceUpdate) {
+        spotifyosa.isApplicationPlaying().then((isPlaying) => {
+            if (previousIsPlayingState != isPlaying) {
+                previousIsPlayingState = isPlaying;
+                tpclient.stateUpdate("state_spotifypg_play", isPlaying);
+            }
+
+            if (shouldUpdateTrackInfoStates) {
+                getTrackInfoStates();
+            }
+        });
+    }
+}
+
+function getTrackInfoStates() {
+    spotifyosa.getTrackTitle().then((result) => {
+        currentTrackTitle = result;
+        if (currentTrackTitle != "" && previousIsPlayingState == "playing") {
+            spotifyosa.getTrackAlbum().then((result) => {
+                currentTrackAlbum = result;
+
+                spotifyosa.getTrackArtist().then((result) => {
+                    currentTrackArtist = result;
+
+                    spotifyosa.getTrackAlbumArtist().then((result) => {
+                        currentTrackAlbumArtist = result;
+
+                        spotifyosa.getTrackDuration().then((result) => {
+                            currentTrackDuration = result;
+
+                            spotifyosa.getTrackId().then((result) => {
+                                currentTrackId = result;
+
+                                spotifyosa.getTrackArtworkUrl().then((result) => {
+                                    currentTrackArtworkUrl = result;
+
+                                    sendTrackInfoStates(currentTrackTitle, currentTrackAlbum, currentTrackArtist, currentTrackAlbumArtist, currentTrackDuration, currentTrackId, currentTrackArtworkUrl);
+                                });
+                            });
+                        });
+                    });
+                });
+            });
+        }
+        else if (currentTrackTitle == "Advertisement") {
+            sendTrackInfoStates("Advertisement", "", "", "", "", "", "");
+        }
+        else {
+            sendTrackInfoStates("", "", "", "", "", "", "");
+        }
+    });
+}
+
+function sendTrackInfoStates(trackTitle, trackAlbum, trackArtist, trackAlbumArtist, trackDuration, trackId, trackArtworkUrl) {
+    let states = [
+        { id: 'state_spotifypg_track_title', value: trackTitle },
+        { id: 'state_spotifypg_track_album', value: trackAlbum },
+        { id: 'state_spotifypg_track_artist', value: trackArtist },
+        { id: 'state_spotifypg_track_album_artist', value: trackAlbumArtist },
+        { id: 'state_spotifypg_track_duration', value: trackDuration },
+        { id: 'state_spotifypg_track_id', value: trackId },
+        { id: 'state_spotifypg_track_artwork', value: trackArtworkUrl },
+    ];
+    tpclient.stateUpdateMany(states);
+}
+
 function sendDefaultStates() {
     let states = [
         { id: "state_spotifypg_visible", value: "false" },
+        { id: "state_spotifypg_play", value: "paused" },
     ];
     tpclient.stateUpdateMany(states);
 }