+++ src/osascript.js
@@ -4,6 +4,7 @@
 const osascript = require('node-osascript');
 const Promise = require('bluebird');
 const execute = Promise.promisify(osascript.execute);
+const moment = require('moment');
 
 
 function isApplicationOpen() {
@@ -30,6 +31,62 @@ function showApplication() {
     return execute('tell application "System Events"\n set visible of application process "Spotify" to true\n end tell');
 }
 
+function isApplicationPlaying() {
+    return execute('tell application "Spotify" to get player state as string');
+}
+
+function playCurrentTrack() {
+    return execute('tell application "Spotify" to play');
+}
+
+function pauseCurrentTrack() {
+    return execute('tell application "Spotify" to pause');
+}
+
+function playpauseCurrentTrack() {
+    return execute('tell application "Spotify" to playpause');
+}
+
+function playNextTrack() {
+    return execute('tell application "Spotify" to next track');
+}
+
+function playPreviousTrack() {
+    return execute('tell application "Spotify"\n if player position is greater than 3 then\n previous track\n previous track\n else\n previous track\n end if\n end tell');
+}
+
+function getTrackTitle() {
+    return execute('tell application "Spotify" to name of current track as string');
+}
+
+function getTrackAlbum() {
+    return execute('tell application "Spotify" to album of current track as string');
+}
+
+function getTrackArtist() {
+    return execute('tell application "Spotify" to artist of current track as string');
+}
+
+function getTrackAlbumArtist() {
+    return execute('tell application "Spotify" to album artist of current track as string');
+}
+
+function getTrackDuration() {
+    return execute('tell application "Spotify" to duration of current track').then((duration) => {
+        return moment.duration(duration, 'milliseconds').asSeconds();
+    });
+}
+
+function getTrackId() {
+    return execute('tell application "Spotify" to ID of current track as string').then((id) => {
+        return id.match(/(.*):(.*)/)[2];
+    });
+}
+
+function getTrackArtworkUrl() {
+    return execute('tell application "Spotify" to artwork url of current track as string');
+}
+
 function quitApplication() {
     return execute('tell application "Spotify" to quit');
 }
@@ -41,5 +98,18 @@ module.exports = {
     isApplicationVisible,
     hideApplication,
     showApplication,
+    isApplicationPlaying,
+    playCurrentTrack,
+    pauseCurrentTrack,
+    playpauseCurrentTrack,
+    playNextTrack,
+    playPreviousTrack,
+    getTrackTitle,
+    getTrackAlbum,
+    getTrackArtist,
+    getTrackAlbumArtist,
+    getTrackDuration,
+    getTrackId,
+    getTrackArtworkUrl,
     quitApplication,
 };